Rename products package to store_product & Relocate BaseHandler class
