+++ app/bot.py
@@ -1,6 +1,6 @@
 from google.cloud import dialogflow
-from app.products.product_info import *
-from app.products.store_info import *
+from app.store_product.product_info import *
+from app.store_product.store_info import *
 from app.concerns.other_concern import *
 import random
 import sys
